+++ pkg/aws/v1alpha3/awscluster/validate_awscluster_test.go
@@ -65,3 +65,47 @@ func TestAWSCNIPrefix(t *testing.T) {
 		})
 	}
 }
+
+func Test_AWSClusterAlreadyExists(t *testing.T) {
+	testCases := []struct {
+		name  string
+		valid bool
+	}{
+		{
+			name:  "case 0: cluster does not exists",
+			valid: true,
+		},
+		{
+			name:  "case 1: cluster already exists",
+			valid: false,
+		},
+	}
+
+	for i, tc := range testCases {
+		t.Run(strconv.Itoa(i), func(t *testing.T) {
+			v := &Validator{
+				k8sClient: unittest.FakeK8sClient(),
+				logger:    microloggertest.New(),
+			}
+			awsCluster := unittest.DefaultAWSCluster()
+
+			if !tc.valid {
+				// create the same AWS cluster in giantswarm namespace
+				awsCluster.Namespace = "giantswarm"
+				err := v.k8sClient.CtrlClient().Create(context.TODO(), awsCluster)
+				if err != nil {
+					t.Fatalf("unexpected error %v", err)
+				}
+			}
+
+			// check if the result is as expected
+			err := v.AWSClusterExists(awsCluster)
+			if tc.valid && err != nil {
+				t.Fatalf("unexpected error %v", err)
+			}
+			if !tc.valid && err == nil {
+				t.Fatalf("expected error but returned %v", err)
+			}
+		})
+	}
+}